recognition/detection/image-recognition/requirements.txt
toml
torch
torchvision
colorama
ultralytics
pyzmq